requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
pymupdf>=1.23.0

